% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/graphlm.R
\name{by1var}
\alias{by1var}
\title{Diagnostic plot of one variable of a linear model}
\usage{
by1var(oldLm, var, thin = 1, breakupby = FALSE, adjustedData = F,
  rawData = F, connections = F, loess = F, line = F, rug = F)
}
\arguments{
\item{oldLm}{Linear model to diagnose}

\item{var}{string giving name of variable to plot on x. Must be an
independent variable of the given linear model.}

\item{thin}{Linear model to diagnose

If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.}
}
\description{
This function takes a linear model and creates a plot where the x axis is the given
variable and the y axis shows the raw data (in gray by default) and the data "adjusted for"
all other variables in the model (in red by default). The adjusted data are the residuals,
recentered at the average value.
}

