% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/graphlm.R
\name{by1var.seq}
\alias{by1var.seq}
\title{Diagnostic plot of all variables of a linear model}
\usage{
by1var.seq(theModel, ...)
}
\arguments{
\item{theModel}{Linear model to diagnose}

\item{var}{string giving name of variable to plot on x. Must be an
independent variable of the given linear model.}

\item{thin}{How much to thin the data (to avoid overdense plots)}

\item{breakupby}{A categorical variable on which to break into multiple plots}

\item{adjustedData}{ggplot graphical parameters for the "red" dots}

\item{rawData}{ggplot graphical parameters for the "gray" dots}

\item{adjustedData}{ggplot graphical parameters for the "red" dots}

\item{loess}{ggplot graphical parameters for the fitted loess}

\item{line}{ggplot graphical parameters for the linear fit line}

\item{rug}{ggplot graphical parameters for marginal rug plot}
}
\description{
This function takes a linear model and creates a plots for each variable.
}

