% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/graphlm.R
\name{graphml

This package makes improved diagnostic plots for linear models.}
\alias{graphml

This package makes improved diagnostic plots for linear models.}
\title{Better diagnostic plots for linear models}
\description{
Better diagnostic plots for linear models
}

